I have fought well

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,12 +8,12 @@
 int main(void)
 {
 	int i = 0;
-	float previous = 0, next = 1, current;
+	long double previous = 0, next = 1, current;
 
 	while (i < 98)
 	{
 		current = previous + next;
-		printf("%.0f", current);
+		printf("%.0Lf", current);
 		previous = next;
 		next = current;
 		/* Dont ask me how I got this number, I am just playing */
